+++ bot/listener.js
